% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{All}
\alias{All}
\title{Fish, isolation and survey identities combined for the second and third surveys}
\format{
Factor w/ 12 levels and 44 observations
}
\source{
{Experimental Data}
}
\usage{
All
}
\description{
Fish, isolation and survey identities combined for the second and third surveys
}
\keyword{datasets}
