% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta_deviation_function.R
\name{beta_deviation}
\alias{beta_deviation}
\title{Computing Beta-Deviation}
\usage{
beta_deviation(
  com,
  method = "quasiswap",
  dist = "bray",
  fixedmar = "both",
  shuffle = "both",
  group,
  strata = NULL,
  mtype = "count",
  times = 1000,
  burnin = 0,
  thin = 1,
  transform = NULL,
  seed = NULL
)
}
\arguments{
\item{com}{The community data. A species by site matrix containing abundance or presence/absence data.}

\item{method}{The chosen null model algorithm. See help file from the \code{permatswap} function in package \code{vegan}. Defaut is set to "quasiswap"}

\item{dist}{The beta-diversity metric to be used. Default is set to bray-curtis beta dissimilarity.}

\item{fixedmar}{character, stating which of the row/column sums should be preserved ("none", "rows", "columns", "both"). Defaut is to "both".}

\item{shuffle}{Character, indicating whether individuals ("ind"), samples ("samp") or both ("both") should be shuffled, see details.}

\item{group}{a grouping factor to compute community variability}

\item{strata}{Numeric vector or factor with length same as nrow(m) for grouping rows within strata for restricted permutations. Unique values or levels are used. It is usefull to keep gamma-diversity within treatments constant. Defaut is set to "NULL".}

\item{mtype}{Matrix data type, either "count" for count data, or "prab" for presence-absence type incidence data. Defaut is set to "count".}

\item{times}{Number of permuted matrices. Defaut is set to 1000.}

\item{thin}{Number of discarded permuted matrices between two evaluations in sequential ("swap", "tswap") methods. Defaut is set to 1.}

\item{transform}{A character. Should the community matrix values be transformed? Default is set to "NULL". Options here are the same as for the \code{decostand} function from \code{vegan}.}

\item{seed}{Do you wish set a seed? Defaut is to "NULL"}

\item{burning}{Number of null communities discarded before proper analysis in sequential ("swap", "tswap") methods. Defaut is set to 0.}
}
\description{
This function computes observed beta-diversity (community variability), expected beta-diversity and beta-deviation for a given grouping factor. It strongly rely on functions from package \code{vegan}
}
