% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta_deviation_siqueira_et_al_2019_function.R
\name{beta_deviation_siqueira_et_al_2019}
\alias{beta_deviation_siqueira_et_al_2019}
\title{Computing Beta-Deviation}
\usage{
beta_deviation_siqueira_et_al_2019(
  com,
  dist = "bray",
  keep.gamma = T,
  group,
  times = 1000,
  seed = NULL,
  transform = NULL
)
}
\arguments{
\item{com}{The community data. A species by site matrix containing abundance or presence/absence data.}

\item{dist}{The beta-diversity metric to be used. Default is set to bray-curtis beta dissimilarity.}

\item{keep.gamma}{Should permutations be restrict to each treatment?}

\item{group}{a grouping factor to compute community variability}

\item{times}{Number of permuted matrices. Default is set to 1000.}

\item{seed}{Do you wish set a seed? Default is to "NULL"}

\item{transform}{A character. Should the community matrix values be transformed? Default is set to "NULL". Options here are the same as for the \code{decostand} function from \code{vegan}.}
}
\description{
This function computes observed beta-diversity (community variability), expected beta-diversity and beta-deviation for a given grouping factor. It is a changed version of the "beta_deviation" function to use the same null model as used in "Siqueira et al. 2020"
}
\references{
#'Siqueira, Tadeu, Victor S. Saito, Luis M. Bini, Adriano S. Melo, Danielle K. Petsch, Victor L. Landeiro, Kimmo T. Tolonen, Jenny Jyrkänkallio‐Mikkola, Janne Soininen, and Jani Heino. “Community Size Can Affect the Signals of Ecological Drift and Niche Selection on Biodiversity.” Ecology 101, no. 6 (2020): e03014. \url{https://doi.org/10.1002/ecy.3014}.
}
