% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{com_SS1_predators}
\alias{com_SS1_predators}
\title{Taxa abundances of predatory insects for the First survey}
\format{
A data frame with 24 rows or 'observations' and 18 variables or 'taxa':
}
\source{
{Experimental Data}
}
\usage{
com_SS1_predators
}
\description{
A dataset containing the abundances of predatory taxa that were present in the first and third surveys.
}
\keyword{datasets}
