% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{com_SS2_SS3}
\alias{com_SS2_SS3}
\title{Taxa abundances for the second and third surveys}
\format{
A data frame with 44 rows or 'observations' and 24 variables or 'taxa':
}
\source{
{Experimental Data}
}
\usage{
com_SS2_SS3
}
\description{
A dataset containing the abundances of the taxa that were present in more than three observations in the second and third surveys.
}
\keyword{datasets}
