% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{com_SS2_SS3_abundance}
\alias{com_SS2_SS3_abundance}
\title{Abundances for the second and third surveys}
\format{
Numerical vector w/ with 44 observations
}
\source{
{Experimental Data}
}
\usage{
com_SS2_SS3_abundance
}
\description{
Abundances for the second and third surveys
}
\keyword{datasets}
