% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{com_SS2_SS3_non_predators_richness}
\alias{com_SS2_SS3_non_predators_richness}
\title{Richness for non-predatory insects (herbivores and detritivores) at the second and third surveys}
\format{
Numerical vector w/ with 44 observations
}
\source{
{Experimental Data}
}
\usage{
com_SS2_SS3_non_predators_richness
}
\description{
Richness for non-predatory insects (herbivores and detritivores) at the second and third surveys
}
\keyword{datasets}
