% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{com_SS2_SS3_predators}
\alias{com_SS2_SS3_predators}
\title{Taxa abundances for predatory insects at the second and third surveys}
\format{
A data frame with 44 rows or 'observations' and 8 variables or 'taxa':
}
\source{
{Experimental Data}
}
\usage{
com_SS2_SS3_predators
}
\description{
A dataset containing the abundances of the predatory insects that were present in the second and third surveys.
}
\keyword{datasets}
