% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{com_SS3}
\alias{com_SS3}
\title{Taxa abundances for the Third survey}
\format{
A data frame with 20 rows or 'observations' and 29 variables or 'taxa':
}
\source{
{Experimental Data}
}
\usage{
com_SS3
}
\description{
A dataset containing the abundances of the taxa that were present in the second and third surveys.
}
\keyword{datasets}
