% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{com_SS3_non_predators}
\alias{com_SS3_non_predators}
\title{Taxa abundances of non-predatory insects for the Third survey}
\format{
A data frame with 20 rows or 'observations' and 29 variables or 'taxa':
}
\source{
{Experimental Data}
}
\usage{
com_SS3_non_predators
}
\description{
A dataset containing the abundances of non-predatory taxa that were present in the third and third surveys.
}
\keyword{datasets}
