% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{env_data_all}
\alias{env_data_all}
\title{Environmental variables in all surveys}
\format{
A data frame with 72 rows or 'observations' and 10 variables:
\describe{
\item{pond}{Pond IDs}
\item{survey}{One of the three surveys}
\item{depth}{Average depth (cm) of each pond}
\item{conductivity}{Water conductivity of each pond in each survey}
\item{ph}{Water pH in each pond in each survey}
\item{do}{Dissolved Oxygen (mg/L) in each pond in each survey}
\item{turbidity}{Water turbidity (NTU) in each pond in each survey}
\item{clorophylla_a}{Clorophylla-a values in each pond in each survey}
\item{phycocyanin}{Phycocyanin values in each pond in each survey}
\item{canopy}{Average canopy cover value (\%) in each pond in each survey}
}
}
\source{
{Experimental Data}
}
\usage{
env_data_all
}
\description{
Environmental variables in all surveys
}
\keyword{datasets}
