% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{env_data_SS2_SS3}
\alias{env_data_SS2_SS3}
\title{Environmental variables in the second and third surveys}
\format{
A data frame with 20 rows or 'observations' and 8 variables:
\describe{
\item{depth}{Average depth (cm) of each pond}
\item{conductivity}{Water conductivity of each pond}
\item{ph}{Water pH in each pond}
\item{do}{Dissolved Oxygen (mg/L) in each pond}
\item{turbidity}{Water turbidity (NTU) in each pond}
\item{clorophylla_a}{Clorophylla-a values in each pond}
\item{phycocyanin}{Phycocyanin values in each pond}
\item{canopy}{Average canopy cover value (\%) in each pond}
}
}
\source{
{Experimental Data}
}
\usage{
env_data_SS2_SS3
}
\description{
Environmental variables in the second and third surveys
}
\keyword{datasets}
