% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fish_isolation_SS3}
\alias{fish_isolation_SS3}
\title{Isolation and fish treatment identities for the Third survey}
\format{
Factor w/ 3 levels: "030","120","480", and 20 observations
}
\source{
{Experimental Data}
}
\usage{
fish_isolation_SS3
}
\description{
Isolation and fish treatment identities for the Third survey
}
\keyword{datasets}
