% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fish_SS1}
\alias{fish_SS1}
\title{Fish treatment identities for the First survey}
\format{
Factor w/ 2 levels: "absent","present", and 24 observations
}
\source{
{Experimental Data}
}
\usage{
fish_SS1
}
\description{
Fish treatment identities for the First survey
}
\keyword{datasets}
