% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fish_SS2_SS3}
\alias{fish_SS2_SS3}
\title{Fish treatment identities for the second and third surveys}
\format{
Factor w/ 2 levels: "absent","present", and 44 observations
}
\source{
{Experimental Data}
}
\usage{
fish_SS2_SS3
}
\description{
Fish treatment identities for the second and third surveys
}
\keyword{datasets}
