% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fish_SS3}
\alias{fish_SS3}
\title{Fish treatment identities for the Third survey}
\format{
Factor w/ 2 levels: "absent","present", and 20 observations
}
\source{
{Experimental Data}
}
\usage{
fish_SS3
}
\description{
Fish treatment identities for the Third survey
}
\keyword{datasets}
