% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ID_SS2_SS3}
\alias{ID_SS2_SS3}
\title{Pond IDs for the second and third surveys}
\format{
Factor w/ 24 levels and 44 observations
}
\source{
{Experimental Data}
}
\usage{
ID_SS2_SS3
}
\description{
Pond IDs for the second and third surveys
}
\keyword{datasets}
