% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{isolation_SS2_SS3}
\alias{isolation_SS2_SS3}
\title{Isolation identities for the second and third surveys}
\format{
Factor w/ 3 levels: "30","120", "480" and 44 observations
}
\source{
{Experimental Data}
}
\usage{
isolation_SS2_SS3
}
\description{
Isolation identities for the second and third surveys
}
\keyword{datasets}
