% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/null_assembly_function.R
\name{null_assembly}
\alias{null_assembly}
\title{Null assembly}
\usage{
null_assembly(test, species.col = "species", site.col = "site")
}
\description{
This function is to run the null model used by "Siqueira et al. 2020"
}
\references{
#'Siqueira, Tadeu, Victor S. Saito, Luis M. Bini, Adriano S. Melo, Danielle K. Petsch, Victor L. Landeiro, Kimmo T. Tolonen, Jenny Jyrkänkallio‐Mikkola, Janne Soininen, and Jani Heino. “Community Size Can Affect the Signals of Ecological Drift and Niche Selection on Biodiversity.” Ecology 101, no. 6 (2020): e03014. \url{https://doi.org/10.1002/ecy.3014}.
}
