% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{SS_SS2_SS3}
\alias{SS_SS2_SS3}
\title{Sampling Survey identities for the second and third surveys}
\format{
Factor w/ 2 levels: "1","2", and 44 observations
}
\source{
{Experimental Data}
}
\usage{
SS_SS2_SS3
}
\description{
Sampling Survey identities for the second and third surveys
}
\keyword{datasets}
