% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Trait_SS2}
\alias{Trait_SS2}
\title{Traits and other relevant information for the second survey only}
\format{
A data frame with 30 rows or 'observations' and 16 variables:
\describe{
\item{species}{Factor variable: Ids of the sampled taxa}
\item{trophic}{Factor variable: Trophic level of the sampled taxa}
\item{microhabitat}{Factor variable: Microhabitat level of the sampled taxa}
\item{volume}{Numeric variable: Volume of the largest individual of each sampled taxa}
\item{mass}{Numeric variable: Wet mass of the largest individual of each sampled taxa}
\item{volume_log}{Numeric variable: Log transformed volume of the largest individual of each sampled taxa}
\item{family}{Character variable: Family of each sampled taxa}
\item{order}{Character variable:  Order of each sampled taxa}
\item{total_ab}{Numeric variable: Total abundance of each sampled taxa considering all surveys}
\item{total_ab_log}{Numeric variable: Log transformed total abundance of each sampled taxa considering all surveys}
\item{total_ab_SS1}{Numeric variable: Total abundance of each sampled taxa considering the first survey}
\item{total_ab_SS2}{Numeric variable: Total abundance of each sampled taxa considering the second survey}
\item{total_ab_SS3}{Numeric variable: Total abundance of each sampled taxa considering the third survey}
\item{total_ab_SS1_log}{Numeric variable: Log transformed total abundance of each sampled taxa considering the first survey}
\item{total_ab_SS2_log}{Numeric variable: Log transformed total abundance of each sampled taxa considering the second survey}
\item{total_ab_SS3_log}{Numeric variable: Log transformed total abundance of each sampled taxa considering the third survey}
}
}
\source{
{Experimental Data}
}
\usage{
Trait_SS2
}
\description{
Traits and other relevant information about each taxa sampled
}
\keyword{datasets}
